+++ src/main/java/subway/path/caculator/FareCalculator.java
@@ -2,6 +2,7 @@ package subway.path.util;
 
 import subway.line.domain.Line;
 import subway.line.domain.Section;
+import subway.path.domain.ComplimentaryAge;
 
 import java.util.Comparator;
 import java.util.List;
