+++ src/main/java/subway/path/domain/ComplimentaryAge.java
@@ -1,4 +1,4 @@
-package subway.path.util;
+package subway.path.domain;
 
 import subway.exception.WrongInputDataException;
