+++ src/test/java/subway/path/caculator/FareTest.java
@@ -1,6 +1,5 @@
 package subway.path.util;
 
-import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -9,6 +8,7 @@ import subway.line.domain.Line;
 import subway.line.domain.Section;
 import subway.line.domain.Sections;
 import subway.member.domain.LoginMember;
+import subway.path.domain.ComplimentaryAge;
 import subway.station.domain.Station;
 
 import java.util.Arrays;
